# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Welcome to E-Kart, where you can get & order all the games you want. Register and Enjoy our services.",
    "version": "1.0.0",
    "title": "Swagger Gamestore",
    "contact": {
      "email": "mohit_soni87@ymail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Game",
      "description": "Everything about your Games",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "store",
      "description": "Access to Gamestore orders"
    },
    {
      "name": "user",
      "description": "Operations about user",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/registeruser": {
      "post": {
        "tags": [
          "RegisterUser"
        ],
        "summary": "Add a new user",
        "description": "",
        "operationId": "addGame",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Username",
            "in": "query",
            "description": "Username",
            "required": true,
            "type": "string"
          },
          {
            "name": "Password",
            "in": "query",
            "description": "Password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Login",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Username",
            "in": "query",
            "description": "Username",
            "required": true,
            "type": "string"
          },
          {
            "name": "Password",
            "in": "query",
            "description": "Password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Logout"
        ],
        "summary": "Logout",
        "description": "Multiple status values can be provided with comma separated strings",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Cart Items",
        "description": "Multiple status values can be provided with comma separated strings",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/getall": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get all available Games from all studios",
        "description": "Multiple status values can be provided with comma separated strings",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/gamedetail": {
      "post": {
        "tags": [
          "GameDetail"
        ],
        "summary": "Specific Game detail",
        "description": "",
        "operationId": "uploadFile",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "GameName",
            "in": "query",
            "description": "ID of Game to update",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Fifa 18",
                "Cricket 17",
                "FarCry New Dawn",
                "Assassin's Creed Origins",
                "Assassin's Creed Syndicate",
                "Assassin's Creed II"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/addtoCart": {
      "post": {
        "tags": [
          "AddToCart"
        ],
        "summary": "Add a game to the Cart",
        "description": "",
        "operationId": "CartFile",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "GameName",
            "in": "query",
            "description": "Game Name",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Fifa 18",
                "Cricket 17",
                "FarCry New Dawn",
                "Assassin's Creed Origins",
                "Assassin's Creed Syndicate",
                "Assassin's Creed II"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/order/{gameName}": {
      "post": {
        "tags": [
          "OrderGame"
        ],
        "summary": "Order a Game",
        "description": "",
        "operationId": "Order",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "gameName",
            "in": "path",
            "description": "GameName",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Fifa 18",
                "Cricket 17",
                "FarCry New Dawn",
                "Assassin's Creed Origins",
                "Assassin's Creed Syndicate",
                "Assassin's Creed II"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/orderCart": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "Buy Cart",
        "description": "Multiple status values can be provided with comma separated strings",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Gamestore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://Gamestore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:Games": "modify Games in your account",
        "read:Games": "read your Games"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Game": {
      "type": "object",
      "required": [
        "name",
        "photoUrls"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "doggie"
        },
        "photoUrls": {
          "type": "array",
          "xml": {
            "name": "photoUrl",
            "wrapped": true
          },
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/Tag"
          }
        },
        "status": {
          "type": "string",
          "description": "Game status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      },
      "xml": {
        "name": "Game"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}